[pwm: to review] Desc, Enum: Fix lazyness issue when matching on tags
Quoting Peter:

"We shouldn't ever pattern match on the pattern synonym constructors
for levitated stuff, i.e. the following should not appear on the left
of any program:

>   pattern DONE = CON (PAIR ZE VOID)
>   pattern ARG s d = CON (PAIR (SU ZE) (PAIR s (PAIR d VOID)))
>   pattern ARGF s d = CON (PAIR (SU (SU ZE)) (PAIR s (PAIR d VOID)))
>   pattern IND h d = CON (PAIR (SU (SU (SU ZE))) (PAIR h (PAIR d VOID)))
>   pattern IND1 d = CON (PAIR (SU (SU (SU (SU ZE)))) (PAIR d VOID))

Since matching on these means matching on PAIR, which is forbad.

Instead we should match on the CON, project out the tag and match on
that, and then project out the payload, thus preserving the lazyness
of pairs."