[Need review] Fix the fix of the fix of the labelled Mu bug.
Same bug, 3rd edition.

In the previous edition, I got fed up with labels on anonymous Mu and
said: "we don't label anonymous guys, problem solved". Therefore,
equality and coercion would halfZip 'Nothing' as a label, so there was
no more 'label' comparison. In that particular case... Because if at
least one was labelled, then label comparison would occur: the name of
things started to have an importance. In particular, if two
Descriptions were the same (structurally) but had different names,
they asked for a proof of name equality. The system was kind of
nominal, or I would rather say 'completely broken'.

In this patch, I revert this second edition. eqGreen goes back to 1st
edition: we ignore label, check the structure for equality. Unlike the
1st time, I have not forgotten coerce this time. In theory, now, the
equality is back to a structural version. Label equality is sound, but
not complete. If we fallback on eqGreen or equality, this means we
really have to do the structural comparisons.

I wonder what will be the fix of the fix of the fix of the fix.