Cat.pig: Diag category defined.
Don't read this file thinking that Epigram is ready for CT. This file
is more an S&M experience than anything else.

Morevoer, using the official Epigram, this won't type-check (simpl
hangs). Also, the last line opens a vortex to Jupiter and try to suck
it up on your RAM too.

So, really, this is not for the faint of heart.