Desc@ICFP: fix Hutton's example
Instead of fiddling with two separate finite domains for each types,
we really define a notion of context, environment, and variable
pointing into that.

Addressing the question of a reviewer, the model contains the case of
pairs, mixing any odd type.

A consequence is that we talk about equality. But that should be ok.