Desc@ICFP: justify the adoption of a non-algebraic code for Desc
Reviewer 2 says:

"Fig. 3.  I couldn't figure out why 'indx takes a description argument
and is defined as X x [[D]] X instead of just being a constant
interpreted as X.  (Indeed, this is apparently done in the
implementation, and 'hindx is apparently replaced by pi).Similarly, in
sec. 4.4, it took me a second to see why you embed X as 'Sigma X '1 -
but presumably you could have a 'const description that embeds an
arbitrary X : Set into a Desc (just as you do for IDesc)."

And we said:

"Referee 23B also suggests that we use the "algebraic" style for both
Desc and IDesc. It is a temptation, but we also want to keep the
connection between Desc and familiar datatype declarations by
enforcing their right-nested tuple structure. We shall certainly try
to make the notation and motivation clearer."