Desc@ICFP: (try to) be consistent on spacing, parenthesing, etc.
Reviewer 2 said:

"First, generally, parentheses and spacing are not used consistently,
and I had to guess what the precedence of various operations was too
often (or try to do the type inference in my head).  For example, D* X
is parenthesized in the type/definition of subst but not in apply (on
p. 7)."

I tried to be consistent. But Godel tells me that I cannot garantee my
own consistency, so someone will have to double check.