Desc@ICFP: develop our position for Set : Set
Reviewer 1 says:

"However, the idea of having a type of data type of encodings which
contains its own code is just too close to having a type of all types
to be taken lightly. It doesn't help that the authors "locally adopt
SET : SET, putting presentation before paradox", since this is related
to the key theoretical issue of the paper."

To which we said:

"As presented with Set:Set, our system is clearly open to
nontermination via paradox.  Choosing a particular stratification
discipline is an option but would introduce extra machinery: we shall
investigate the possibility of such a presentation and do so if space
and comprehensibility permit."

And decided to modify the "putting presentation before paradox" to
something saying:
  - We takle one problem at a time: data-types here,
    stratification is another
  - We give a basic picture 
  - There is further work needed to justify the all of it
  - We are working on it