ProofState: merge Movement into Edition
Despite my best will, it is very difficult to distinguish the notion
of "movement" inside the proof context, from the notion of "edition"
of the proof context. This is due to the fact that, in the proof
context, definitions under the cursor can contain news bulletins,
which are (lazy) edition commands.

One could think of putting movement-only commands in "movement" (such
as cursorUp and the like), but this is not really natural. I'm a bit
sad to lose the distinction movement/edition, but it seems
unavoidable.