ProofState.Edition.GetSet: update get/set functions to new names
Taking advantage of the re-naming, I've tried to put some kind of
structure in this hell. The presentation is top-down, with
get/put/remove/replace functions for |ProofContext|, then going deeper
in the bowels of the monster. It is not perfectly regular, as some
functions exists in some context and not others.

The renamings are the following.

* GetSet in |ProofContext|:
** (get|put|replace)Dev: (\1)AboveCursor
** (get|put)DevCadets: (\1)BelowCursor

* GetSet in |AboveCursor|:
** (get|put|replace)DevEntries: (\1)EntriesAbove
** (get|put|remove)DevEntry: (\1)EntryAbove

* GetSet in |Layers|:
** (get|put)Mother: (\1)CurrentEntry
** insertCadet: putNewsBelow

* GetSet in |CurrentEntry|:
** getMotherName: getCurrentName
** getMotherDefinition: getCurrentDefinition
** putMotherScheme: putCurrentScheme

* Get Scopes:
** getGreatAuncles: getGlobalScope
** getAunclesToImpl: getDefinitionsToImpl
** getBoys: getParamsInScope

Moreover, a few functions have moved to a newly created
ProofState.Edition.Navigation file:
** getMotherEntry: getLeaveCurrent
** putMotherEntry: putLeaveCurrent
This file should be extended in the future.