Epitome: fix \label and \ref links
I've adopted some kind of standard scheme for defining '\label's that
prevents name clashes and doesn't require omniscience (very useful
indeed).

So, when declaring a label in An.Haskell.Module on, say, the following section:

\subsubsection{Foo bar}

The label is as follow:

\label{subsubsec:An.Haskell.Module.whatever-you-want}

For figures, equations, and tables, one would write 'fig:', 'eqn:',
and 'tbl:' instead of 'subsubsec'.

When labelling the '\section{}' of a module An.Other.Module, we write:

\label{sec:An.Other.Module}