Move discharge code from ProofState.Interface.Lifting to new file Evidences.Utilities
This file is for code that does not depend on the proof state or display 
language, but is not essential for defining the evidence language or its
computational behaviour (otherwise it would live in Tm or Rules). Some of
the code in those files could probably move to Utilities.

The EXTM/INTM transition code at the end of ProofState.Interface.ProofKit
has also moved here, because it was not ProofState-specific.