PropSimp: only simplify functions from enumerations when the enumeration is canonical
Ultimately we should get rid of this simplification step altogether, but that
will have to wait until we port everything to use idata and the nice induction
principles. In the meantime, we only simplify functions from canonical
enumerations, which makes generic programming on descriptions (e.g. Below.pig)
a lot nicer.