Stop elaborator from labelling inductive data types
Now that we have the data and idata tactics, we do not need to label inductive data types when we elaborate them. This is good, because sometimes the elaborator would attach labels in the wrong places when doing generic programming.

The IDesc elaboration/distillation/pretty-printing code has been extended to support unlabelled |IMu|s, and Desc has been tidied up a little.