Typechecker: evaluate away more of the irritating redexes
The idea is that users cannot write redexes, but that we might generate them internally (e.g. in the propositional simplifier), so it should be safe to simply evaluate them. However, this means we need to have a rather large trusted kernel and makes debugging harder. This patch makes the typechecker look for redexes (index variables or applied lambdas) that might be hiding under closures. 

Someone more knowledgeable than me should review this patch.