simplified rules about semicolons and added braces
I realised that I was getting ; a bit wrong, so I've fixed things.
A semicolon is always a top-level newline in the current layout stack.
To localize the layout stack, use braces.

So

  let f : Set; ...

is *bad*, because the let has no dialogue and the ... is in the place of
a new definition.

But

  let {f : Set; ...}

is *good*, because the ... is now in the right place to stand for the
missing dialogue.

You can also avoid semicolons by writing the *good*

  let f : Set
    ...

but the indentation is important. But

  let f : Set
  ...

is *bad*, because again, the ... is not subordinate to the let.